[SW-2009] No need to search for client ip in rest api mode
